chore: remove deprecated code, improve structure
