+++ lib/features/settings/features/features.dart
@@ -5,5 +5,4 @@ export 'help/help.dart';
 export 'language/language.dart';
 export 'network/network.dart';
 export 'theme_mode/theme_mode.dart';
-export 'call_codecs/call_codecs.dart';
 export 'self_config/self_config.dart';
